+++ src/features/metrics/components/disks/disk.info.tsx
@@ -1,12 +1,12 @@
 import drive from "/drive.png";
 import { memo } from "react";
-import { useNavigate } from "react-router-dom";
 import { useShallow } from "zustand/react/shallow";
 
 import DynamicProgress from "@/components/dynamic-progress";
 import useDisksStore from "@/features/metrics/stores/disk.store";
 import useSystemStoreSelectors from "@/features/metrics/stores/system.store";
 import formatBytes from "@/features/metrics/utils/format-bytes";
+import useRouteHandler from "@/hooks/useRouteHandler";
 import useTheme from "@/hooks/useTheme";
 import { commands, Disk } from "@/lib";
 import hasBytesTextChanged from "@/utils/has-text-changed";
@@ -121,7 +121,7 @@ const DiskActionGroup: React.FC<{ disk: Disk }> = ({ disk }) => {
   const { classes } = useStyles();
   const { isMidnight } = useTheme();
   const setSelectedDisk = useDisksStore.use.setSelectedDisk();
-  const navigate = useNavigate();
+  const { navigateToDynamic } = useRouteHandler();
   const system = useSystemStoreSelectors(useShallow((state) => state.info));
 
   const showDirectory = async () => {
@@ -134,7 +134,7 @@ const DiskActionGroup: React.FC<{ disk: Disk }> = ({ disk }) => {
   const onShowDetailsClick = () => {
     const isWindows = system.os.toLowerCase().includes("windows");
     setSelectedDisk(isWindows ? disk.name : disk.mountPoint);
-    navigate(`/disks/${encodeURIComponent(isWindows ? disk.name : disk.mountPoint)}`);
+    navigateToDynamic("/disks/:id", encodeURIComponent(isWindows ? disk.name : disk.mountPoint));
   };
 
   return (
@@ -153,7 +153,7 @@ const DiskStatsCard: React.FC<DiskInfoProps> = ({ disk }) => {
   const { classes } = useStyles();
 
   return (
-    <Card shadow="xl" p="xs" radius={"md"} withBorder>
+    <Card shadow="xl" p="xs" radius={"md"} withBorder style={{ backgroundColor: "transparent" }}>
       <Card.Section className={classes.section}>
         <DiskDetailsSection disk={disk} />
       </Card.Section>