+++ src/features/metrics/components/networks/networks-transmitted.area-chart.tsx
@@ -44,7 +44,7 @@ const NetworksTransmittedAreaChart: React.FC = ({}) => {
   }, [JSON.stringify(networks)]);
 
   return (
-    <Card style={{ height: "410px" }}>
+    <Card style={{ height: "380px" }}>
       <AreaChart options={chartOptions} />
     </Card>
   );