+++ vite.config.ts
@@ -1,7 +1,8 @@
-import * as path from "path";
-import { defineConfig, optimizeDeps } from "vite";
+import * as path from 'path';
+import { defineConfig } from 'vite';
+import eslint from 'vite-plugin-eslint';
 
-import react from "@vitejs/plugin-react-swc";
+import react from '@vitejs/plugin-react-swc';
 
 // https://vitejs.dev/config/
 export default defineConfig({